HCK-4339: extract similar logic in a fucntion and reuse it
